refactor: #266 TagListStyle로 파일명 수정
